% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_climatology_df.R
\name{as_climatology_df}
\alias{as_climatology_df}
\title{Flatten climatology outputs (list of SpatRaster) to a tidy data.frame}
\usage{
as_climatology_df(input, vars = NULL, xy = TRUE, rename_xy = TRUE)
}
\arguments{
\item{vars}{character vector of element names to extract and join
(defaults to c("sst","climatology","anomaly","hotspots","dhw") present in x)}

\item{x}{list with SpatRaster elements (e.g., x$sst, x$climatology, x$anomaly, x$hotspots, x$dhw)}
}
\value{
tibble with columns: lon, lat, time, and one column per var in \code{vars}
}
\description{
Flatten climatology outputs (list of SpatRaster) to a tidy data.frame
}
