% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.R
\name{background}
\alias{background}
\title{Run Expression in Background with Completion Message}
\usage{
background(
  expr,
  stdout = NULL,
  stderr = NULL,
  overwrite = TRUE,
  envir = .GlobalEnv
)
}
\arguments{
\item{expr}{An expression wrapped in \code{{}} to be evaluated in a background R process.}

\item{stdout}{Optional file to capture standard output. Defaults to NULL.}

\item{stderr}{Optional file to capture standard error. Defaults to NULL.}

\item{wait}{Logical, if TRUE will block until process is complete and print message. Default: TRUE.}
}
\value{
A \code{callr::process} object.
}
\description{
Launches a background R process to run an expression asynchronously.
Displays a message once the process finishes.
}
\examples{
\dontrun{
background({
  Sys.sleep(5)
  writeLines("Done", "output3.txt")
})
}
}
