% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_anomalies.R
\name{calculate_anomalies}
\alias{calculate_anomalies}
\title{Calculate anomalies}
\usage{
calculate_anomalies(sst_file, climatology)
}
\arguments{
\item{sst_file}{sst file}

\item{climatology}{daily climatology}
}
\value{
SST anomalies (terra::rast format)
}
\description{
Function to calculate anomalies

See vignette for further details
}
