% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_daily_climatology.R
\name{calculate_daily_climatology}
\alias{calculate_daily_climatology}
\title{Calculate Daily Climatology}
\usage{
calculate_daily_climatology(sst_file, mm)
}
\arguments{
\item{sst_file}{sst file}

\item{mm}{monthly maximums}
}
\value{
climatology (terra::rast format)
}
\description{
Function to calculate daily climatology

See vignette for further details
}
