% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dhm.R
\name{calculate_dhmi}
\alias{calculate_dhmi}
\title{Calculate Degree Heating Month Index (DHMi)}
\usage{
calculate_dhmi(
  sst_raster,
  timeseries = c("daily", "monthly"),
  accumulation_window = 3,
  timespan = c(12, 5)
)
}
\arguments{
\item{sst_raster}{A \code{SpatRaster} object containing SST data. The raster should include a time dimension.}

\item{accumulation_window}{time window for calculating DHMi (typically 3 or 4 months, see Table S1 in Mason et al 2024 Nature Geoscience)}

\item{timespan}{start and end months for summed DHMi calculations (for example in Austral summer - November to May = c(12,5)}
}
\value{
A \code{SpatRaster} object with the computed metric.
}
\description{
This function calculates Degree Heating Month Index (DHMi) from Lough et al (2018) and is calculated based on either monthly or daily SST data

Citation: Lough et al (2018) Increasing thermal stress for tropical coral reefs: 1871–2017 Scientific Reports 8(6079)
https://www.nature.com/articles/s41598-018-24530-9
}
\examples{
\dontrun{
library(terra)
sst_data <- rast("dhw_5km_12c5_a60a_4147_U1739222110146.nc")
ssta <- calculate_dhmi(sst_data, timeseries="daily", timespan = c(12, 4))
plot(ssta)
}

}
