% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_hotspots.R
\name{calculate_hotspots}
\alias{calculate_hotspots}
\title{Calculate hotspots}
\usage{
calculate_hotspots(mmm, sst_file, anomaly = 1)
}
\arguments{
\item{mmm}{maximum monthly mean}

\item{sst_file}{sst file}
}
\value{
daily hotpsots (terra::rast format)
}
\description{
Function to calculate hotspots

See vignette for further details
}
