% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_maximum_monthly_mean.R
\name{calculate_maximum_monthly_mean}
\alias{calculate_maximum_monthly_mean}
\title{Create Maximum Monthly Mean (MMM)}
\usage{
calculate_maximum_monthly_mean(mm)
}
\arguments{
\item{mm}{monthly mean?}
}
\value{
output list (see above for details)
}
\description{
Function to calculate maximum monthly mean from datasets
}
