% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_metrics.R
\name{calculate_metrics}
\alias{calculate_metrics}
\title{Calculate SST Variability Metrics}
\usage{
calculate_metrics(rast_obj, lyr, metric)
}
\arguments{
\item{rast_obj}{A \code{SpatRaster} object containing SST data. The raster should include a time dimension.}

\item{metric}{A character string specifying the SST metric to compute. Options include:
\describe{
\item{"SSTA"}{Sea Surface Temperature Anomaly (SST deviation from climatology).}
\item{"SST_SD"}{Standard deviation of SST over time.}
\item{"SST_Trend"}{Linear trend of SST over time.}
\item{"SST_Variance"}{Variance of SST values.}
\item{"SST_Seasonality"}{Annual range (max - min SST).}
\item{"SST_Interannual_Amplitude"}{Difference between annual max and min SST.}
\item{"SST_Skewness"}{Skewness of SST distribution.}
}}
}
\value{
A \code{SpatRaster} object with the computed metric.
}
\description{
This function calculates various sea surface temperature (SST) variability metrics from a \code{SpatRaster} object.
The user can specify the desired metric, which is then computed based on time series data.
}
\details{
\itemize{
\item The function ensures that the input raster contains time information.
\item If an invalid metric is supplied, an error is returned.
}
}
\examples{
\dontrun{
library(terra)
sst_data <- rast("dhw_5km_12c5_a60a_4147_U1739222110146.nc")
ssta <- calculate_metrics(sst_data, "SSTA")
plot(ssta)
}

}
