% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_monthly_mean.R
\name{calculate_monthly_mean}
\alias{calculate_monthly_mean}
\title{Create Monthly Mean (MMM)}
\usage{
calculate_monthly_mean(sst_file)
}
\arguments{
\item{sst_file}{sst file}
}
\value{
climatology (terra::rast format)
}
\description{
Function to calculate maximum monthly mean

See vignette for further details
}
