% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_climatology.R
\name{create_climatology}
\alias{create_climatology}
\title{Create climatology}
\usage{
create_climatology(
  sst_file,
  anomaly = 1,
  window = 84,
  quiet = FALSE,
  return = FALSE,
  baa = FALSE,
  save_output = NULL,
  climatology = NULL,
  ...
)
}
\arguments{
\item{sst_file}{input}

\item{window}{number of days for the DHW sum (12 weeks = 84 days default)}

\item{quiet}{verbose - update with messages?}

\item{return}{return output TRUE/FALSE}

\item{baa}{return baa? TRUE/FALSE (speeds up processing)}

\item{save_output}{save output? "folder/filename" format where "folder/filename_sst.tif", "folder/filename_dhw.tif" etc}

\item{climatology}{replace mmm with external climatology? link to .nc, explicit for CRW (i.e. "GBR_ct5km_climatology_v3.1.nc")}
}
\value{
output list (see above for details)
}
\description{
Function to calculate climatology from datasets

See vignette for further details
outputs:
\itemize{
\item output$sst
\item output$anomalies
\item output$mm
\item output$mmm
\item output$dailyclimatology
\item output$hs
\item output$dhw
}

note - calculating BAA is slow
}
\examples{
\dontrun{
output <- create_climatology("crw.nc")
}
}
