% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_CoralTemp.R
\name{download_CoralTemp}
\alias{download_CoralTemp}
\title{Download CoralTemp data}
\usage{
download_CoralTemp(
  url = "https://www.ncei.noaa.gov/thredds-ocean/catalog/crw/5km/v3.1/nc/v1.0/daily",
  start_date,
  end_date,
  dest_dir,
  variable = "hs",
  mc.cores = 1
)
}
\arguments{
\item{url}{one of the NOAA thredds url}

\item{start_date}{end year}

\item{end_date}{end year}

\item{dest_dir}{save file location}

\item{variable}{one of 'sst', 'ssta', 'hs', 'dhw'}

\item{mc.cores}{set the number of cores}
}
\value{
downloaded nc files to specified location
}
\description{
Function to download and save NetCDF files for CoralTemp
\code{download_CoralTemp()} is a function to download NOAA CRW data

If the number of cores is set to >1, the function uses \code{mclapply} to parallel download datasets, else 1 = single downloads.

Notes:
The NOAA Coral Reef Watch (CRW) daily global 5km Sea Surface Temperature (SST) product, also known as CoralTemp,
shows the nighttime ocean temperature measured at the surface. The CoralTemp SST data product was developed from
two, related reanalysis (reprocessed) SST products and a near real-time SST product. Spanning January 1, 1985
to the present, the CoralTemp SST is one of the best and most internally consistent daily global 5km SST products available.
}
\examples{
\dontrun{
download_CoralTemp(1990, 2020, "/users/era5/", variable= "ssta", cores=10)
}
}
