% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ERA5.R
\name{download_ERA5}
\alias{download_ERA5}
\title{Download ERA5 data}
\usage{
download_ERA5(
  start_year = 1981,
  end_year = 2022,
  region = c(-9, 142, -25, 153),
  ecmwfr_key,
  timeout,
  dest_dir
)
}
\arguments{
\item{start_year}{start year}

\item{end_year}{end year}

\item{ecmwfr_key}{required ecmfr key, see notes below}

\item{timeout}{set a timeout for downloading (minutes)}

\item{dest_dir}{save file location}
}
\value{
downloaded nc files to specified location
}
\description{
Function to download and save NetCDF files for OISST
\code{download_ERA5()} is a function to download ERA5 SST data

Notes:
ERA5 is the fifth generation ECMWF atmospheric reanalysis of the global climate covering the period from January 1940 to present.
}
\examples{
\dontrun{
download_ERA5(1990, 2020, key, 60, "/users/era5/")
}
}
