% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_nc_file_CRW.R
\name{download_nc_file_CRW}
\alias{download_nc_file_CRW}
\title{Download nc files}
\usage{
download_nc_file_CRW(date, base_url, dest_dir, variable)
}
\arguments{
\item{date}{date in ""}

\item{base_url}{one of the NOAA thredds url}

\item{dest_dir}{destination dir}
}
\value{
downloaded nc files to specified location
}
\description{
Function to download and save NetCDF files

coraltemp_v3.1_19850101.nc
ct5km_ssta_v3.1_19850101.nc
ct5km_dhw_v3.1_19850325.nc
ct5km_baa_v3.1_19850325.nc
ct5km_hs_v3.1_19850101.nc
}
\examples{
\dontrun{
download_nc_file(date = 20240101,
  base_url = "https://www.ncei.noaa.gov/data/sea-surface-temperature-optimum-interpolation/v2.1/access/avhrr/",
  variable = "ssta",
  dest_dir = "/Users/rof011/Downloads")
}
}
\keyword{internal}
