% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_nc_file_OISST.R
\name{download_nc_file_OISST}
\alias{download_nc_file_OISST}
\title{Download nc files}
\usage{
download_nc_file_OISST(date, base_url, dest_dir)
}
\arguments{
\item{date}{date in ""}

\item{base_url}{one of the NOAA thredds url}

\item{dest_dir}{destination dir}
}
\value{
downloaded nc files to specified location
}
\description{
Function to download and save NetCDF files
}
\examples{
\dontrun{
download_nc_file(date = 20240101,
base_url = "https://www.ncei.noaa.gov/data/sea-surface-temperature-optimum-interpolation/v2.1/access/avhrr/",
dest_dir = "/Users/rof011/Downloads")
}
}
\keyword{internal}
