% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_OISST.R
\name{download_OISST}
\alias{download_OISST}
\title{Download OISST data}
\usage{
download_OISST(
  url = "https://www.ncei.noaa.gov/thredds/fileServer/OisstBase/NetCDF/V2.1/AVHRR",
  start_date,
  end_date,
  dest_dir,
  mc.cores = 1
)
}
\arguments{
\item{url}{one of the NOAA thredds url}

\item{start_date}{end year}

\item{end_date}{end year}

\item{dest_dir}{save file location}

\item{mc.cores}{set the number of cores}
}
\value{
downloaded nc files to specified location
}
\description{
Function to download and save NetCDF files for CoralTemp
\code{download_OISST()} is a function to download OISST v2.1

If the number of cores is set to >1, the function uses \code{mclapply} to parallel download datasets, else 1 = single downloads.

Notes:
The NOAA 1/4° Daily Optimum Interpolation Sea Surface Temperature (OISST) is a long term Climate Data Record that incorporates
observations from different platforms (satellites, ships, buoys and Argo floats) into a regular global grid. The dataset is
interpolated to fill gaps on the grid and create a spatially complete map of sea surface temperature. Satellite and ship observations
are referenced to buoys to compensate for platform differences and sensor biases.
}
\examples{
\dontrun{
download_CoralTemp(1990, 2020, "/users/era5/", variable = "ssta", cores = 10)
}
}
