% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_reefs.R, R/process_ERA5.R
\name{extract_reefs}
\alias{extract_reefs}
\alias{process_ERA5}
\title{Extract reef}
\usage{
extract_reefs(
  input,
  shpfile,
  output = "sf",
  extract_fun = "weighted_mean",
  quiet = FALSE
)

process_ERA5(folder, units = "celsius")
}
\arguments{
\item{input}{input}

\item{shpfile}{location of shpfile mask}

\item{output}{output format, sf default}

\item{extract_fun}{= "weighted_mean" by default}

\item{quiet}{show verbose}

\item{folder}{folder with nc files}

\item{units}{one of "celsius", "kelvin"}
}
\value{
shp file sf with SST details (see above for details)

combined ERA5 rast format
}
\description{
Function to extract sst data for shp file overlay

See vignette for further details

Function to extract sst data for shp file overlay

See vignette for further details
}
\examples{
\dontrun{

noaa_dhw_file = "https://pae-paha.pacioos.hawaii.edu/erddap/griddap/dhw_5km.nc?CRW_SST\%5B(2024-12-20T12:00:00Z):1:(2024-12-25T12:00:00Z)\%5D\%5B(-8.9):1:(-25)\%5D\%5B(141):1:(153.5)\%5D"
noaa_dhw_raster <- rast(noaa_dhw_file) |> flip() |> project("EPSG:4283")
tmp <- extract_reefs(input=noaa_dhw_raster, shpfile=gbr_files)
tmp
}
\dontrun{

}
}
