% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAA_CRW_data.R
\docType{data}
\name{NOAA_CRW_data}
\alias{NOAA_CRW_data}
\title{NOAA_CRW_data}
\format{
dataframe
\describe{
\item{CRW_SST}{Sea Surface Temperature)}
\item{CRW_SSTANOMALY}{Sea Surface Temperature anomaly)}
\item{CRW_DHW}{Degree Heating Weeks}
}
}
\source{
library(rerddap)

NOAA_CRW <- griddap(
datasetx = 'NOAA_DHW',
time = c("2015-06-01", "2017-06-01"),
latitude = c(-14.655, -14.655),
longitude = c(145.405, 145.405),
fmt = "nc"
)

NOAA_CRW_data <- NOAA_CRW$data
}
\usage{
NOAA_CRW_data
}
\description{
subset of NOAA CRW data from rerdapp
}
\examples{
data(NOAA_CRW_data)
summary(NOAA_CRW_data)
}
\keyword{datasets}
