% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2.R
\name{plot2}
\alias{plot2}
\title{Simple XY plot for terra::rast values() comparisons}
\usage{
plot2(x, y)
}
\arguments{
\item{x}{terra::rast(a)}

\item{y}{terra::rast(b)}
}
\value{
base R plot
}
\description{
Helper function to quickly visualise the values of two terra::rast files in a base r xy plot (with abline)
Assumes one var per rast
}
