% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_annual_dhw.R
\name{plot_annual_dhw}
\alias{plot_annual_dhw}
\title{Plot mean monthly maximum for a given gridcell}
\usage{
plot_annual_dhw(input, fixedyears = NULL)
}
\arguments{
\item{input}{rast file}

\item{fixedyears}{range y lims for ggplot}
}
\value{
ggplot of annual DHW scaled Nov-Aug
}
\description{
Plot mean monthly maximum for a given gridcell
from a terra::rast file input
}
