% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mm.R
\name{plot_mm}
\alias{plot_mm}
\title{Plot mean monthly maximum for a given gridcell}
\usage{
plot_mm(input, lon, lat)
}
\arguments{
\item{input}{rast() file}

\item{lon}{vector of longitude}

\item{lat}{vector of latitude}
}
\value{
ggplot of baseline climatology
}
\description{
Plot mean monthly maximum for a given gridcell
from a terra::rast() file input
}
