% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_year.R
\name{plot_year}
\alias{plot_year}
\title{Calculate anomalies}
\usage{
plot_year(sst_file, climatology)
}
\arguments{
\item{sst_file}{sst file}

\item{climatology}{daily climatology}
}
\value{
SST anomalies (terra::rast format)
}
\description{
Function to calculate anomalies

See vignette for further details
}
