% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_OISST.R
\name{process_OISST}
\alias{process_OISST}
\title{Download GBR spatial data}
\usage{
process_OISST(
  input,
  polygon,
  crop = TRUE,
  mask = TRUE,
  downsample = FALSE,
  res = 0.1,
  variable = "sst",
  crs = "EPSG:7844",
  combinedfilename = NULL
)
}
\arguments{
\item{input}{input folder}

\item{polygon}{polygon for crop/masl}

\item{crop}{TRUE/FALSE}

\item{mask}{TRUE/FALSE}

\item{downsample}{TRUE/FALSE}

\item{res}{resolution for downsamlpling}

\item{variable}{redundant?}

\item{crs}{change the CRS if needed (EPSG:4283 as default)}

\item{combinedfilename}{output file path, should be .rds}
}
\value{
terra::rast
}
\description{
function to download the GBR shape files (14.1mb in size, shp file format) via eReefs

Notes:
There are several versions of the GBR reefs shape file. This version is downloaded via the eAtlas
Website and includes reefs from the Torres Strait. Default CRS is GDA94 (EPSG:4283)
}
\examples{
\dontrun{

GBR_hull <- download_gbr_spatial(return="hull", crs = "EPSG:7844")

process_OISST(input = "/Volumes/Extreme_SSD/dhw/global",
                  polygon = GBR_hull, crs = "EPSG:7844",
                  combinedfilename = "/Volumes/Extreme_SSD/dhw/GBR_OISST_full.rds",
                  crop=TRUE, mask=TRUE, downsample=FALSE)


rast(unwrap("/Volumes/Extreme_SSD/dhw/GBR_CoralTemp_full.rds"))
rast(unwrap("/Volumes/Extreme_SSD/dhw/GBR_CoralTemp_full.rds"))[[1]] |> plot()

}

\dontrun{

process_CoralTemp(input = "/Volumes/Extreme_SSD/dhw/coraltempdhw",
                  polygon = GBR_hull, crs = "EPSG:7844",
                  combinedfilename = "/Volumes/Extreme_SSD/dhw/GBR_CoralTempDHW_full.rds",
                  crop=TRUE, mask=TRUE, downsample=FALSE)


rast(unwrap("/Volumes/Extreme_SSD/dhw/GBR_CoralTempDHW_full.rds"))[[1]] |> plot()
rast(unwrap("/Volumes/Extreme_SSD/dhw/GBR_CoralTempDHW_full.rds"))

}
}
