% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_download_OISST.R
\name{remote_download_OISST}
\alias{remote_download_OISST}
\title{Download Latest OISST and process outputs}
\usage{
remote_download_OISST(folder = ".", reefs = GBR_hull, quiet = FALSE)
}
\arguments{
\item{folder}{Character. Local directory where files will be saved.}

\item{reefs}{shp file input, use dhw::download_gbr_spatial(return = "hull", crs = "EPSG:7844")}

\item{dataset}{Character. One of \code{"sst"}, \code{"ssta"}, \code{"dhw"}, \code{"hs"}, \code{"baa"}, or \code{"all"}.}
}
\value{
Downloads the latest file(s) to the specified folder.
}
\description{
Downloads the most recent \code{.nc} file for OISST
converts to DHW
}
\examples{
\dontrun{
GBR_hull = download_gbr_spatial(return = "hull", crs = "EPSG:7844")
remote_download_CoralTemp(dataset = "sst", folder = "data/", reefs = GBR_hull)
remote_download_CoralTemp(dataset = "all", folder = "data/", reefs = GBR_hull)
}
}
