% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_polygon_from_bbox.R
\name{st_polygon_from_bbox}
\alias{st_polygon_from_bbox}
\title{Create sf polygon from bbox}
\usage{
st_polygon_from_bbox(bbox, output_crs)
}
\arguments{
\item{output_crs}{desired output CRS (integer or string)}

\item{boundaries}{in list: c(xmin, ymin, xmax, ymax)}
}
\value{
sf polygon
}
\description{
Function to convert a bounding box to a polygon with correct CRS.
}
