% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dhw.R
\name{calculate_dhw}
\alias{calculate_dhw}
\title{Calculate DHW}
\usage{
calculate_dhw(hs, anomaly = 1, window = 84)
}
\arguments{
\item{hs}{SpatRaster of hotspots.}

\item{anomaly}{numeric threshold for hotspots (default = 1). If <1, no thresholding is applied.}

\item{window}{number of days to sum hotspots, default = 84 (12 weeks).}
}
\value{
SpatRaster of Degree Heating Weeks.
}
\description{
Function to calculate Degree Heating Weeks (DHW).

The function computes the Degree Heating Weeks (DHW) metric, which accumulates
heat stress over a specified rolling window (defaulting to 84 days). If
\code{anomaly >= 1}, only daily hotspot values greater than or equal to that
threshold are included. If \code{anomaly < 1}, \emph{all} non-NA hotspot values are
included without thresholding.
}
