% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_monthly_mean.R
\name{calculate_monthly_mean}
\alias{calculate_monthly_mean}
\title{Create Monthly Mean (MMM)}
\usage{
calculate_monthly_mean(sst_file, midpoint = 1988.2857, return = "predict")
}
\arguments{
\item{sst_file}{SpatRaster of SST data.}

\item{return}{Type of output: "predict", "slope", or "intercept".}
}
\value{
Climatology as a terra::rast object.
}
\description{
Function to calculate maximum monthly mean or trends.

See vignette for further details.
}
