% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_climatology.R
\name{create_climatology}
\alias{create_climatology}
\title{Create climatology}
\usage{
create_climatology(sst_file, window = 84, quiet = FALSE)
}
\arguments{
\item{sst_file}{input}

\item{window}{number of days for the DHW sum (12 weeks = 84 days default)}

\item{quiet}{verbose - update with messages?}
}
\value{
output list (see above for details)
}
\description{
Function to calculate climatology from datasets

See vignette for further details
outputs:
\itemize{
\item output$sst
\item output$anomalies
\item output$mm
\item output$mmm
\item output$dailyclimatology
\item output$hs
\item output$dhw
}
}
\examples{
\dontrun{
output <- create_climatology("crw.nc")
}
}
