% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_CoralTemp.R
\name{download_CoralTemp}
\alias{download_CoralTemp}
\title{Download CoralTemp data}
\usage{
download_CoralTemp(
  url = "https://www.ncei.noaa.gov/thredds-ocean/fileServer/crw/5km/v3.1/nc/v1.0/daily/",
  start_date = NULL,
  end_date = NULL,
  dates = NULL,
  dest_dir,
  variable = "sst",
  mc.cores = 1,
  quiet = TRUE
)
}
\arguments{
\item{url}{NOAA THREDDS server URL. Default is the CoralTemp data server.}

\item{start_date}{Start date in "YYYY-MM-DD" format.}

\item{end_date}{End date in "YYYY-MM-DD" format.}

\item{dates}{Vector of dates as an alternative to start_date and end_date for non sequential timeseries}

\item{dest_dir}{Directory where NetCDF files should be saved.}

\item{variable}{Data type: 'sst', 'dhw', 'ssta', or 'hs'.}

\item{mc.cores}{Number of cores for parallel downloads.}

\item{quiet}{show verbose? TRUE by default}
}
\value{
Saves NetCDF files in the specified destination folder.
}
\description{
Downloads and saves NOAA CoralTemp NetCDF files from the specified start and end dates.

The CoralTemp dataset provides daily global 5km Sea Surface Temperature (SST) data,
including anomalies and degree heating weeks, spanning from January 1, 1985.

URL links current 4th Jan:
https://www.ncei.noaa.gov/thredds-ocean/catalog/crw/5km/v3.1/nc/v1.0/daily/sst/1985/catalog.html?dataset=crw/5km/v3.1/nc/v1.0/daily/sst/1985/coraltemp_v3.1_19850101.nc
https://www.ncei.noaa.gov/thredds-ocean/fileServer/crw/5km/v3.1/nc/v1.0/daily/sst/1985/coraltemp_v3.1_19850101.nc

See: https://coralreefwatch.noaa.gov/product/5km/index_5km_sst.php
}
\examples{
\dontrun{

download_CoralTemp(url = "https://www.ncei.noaa.gov/thredds-ocean/fileServer/crw/5km/v3.1/nc/v1.0/daily/",
                   start_date = "1990-01-01",
                   end_date = "1990-01-02",
                   dest_dir = "/Volumes/Extreme_SSD/dhw/CRW/2024/",
                   variable = "sst",
                   mc.cores = 1)
}
}
