% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_OISST.R
\name{download_OISST}
\alias{download_OISST}
\title{Download OISST data}
\usage{
download_OISST(
  url =
    "https://www.ncei.noaa.gov/data/sea-surface-temperature-optimum-interpolation/v2.1/access/avhrr/",
  start_date = NULL,
  end_date = NULL,
  dates = NULL,
  dest_dir,
  mc.cores = 1
)
}
\arguments{
\item{url}{one of the NOAA thredds url}

\item{start_date}{end year}

\item{end_date}{end year}

\item{dates}{Vector of dates as an alternative to start_date and end_date for non sequential timeseries}

\item{dest_dir}{save file location}

\item{mc.cores}{set the number of cores}
}
\value{
downloaded nc files to specified location
}
\description{
Function to download and save NetCDF files for CoralTemp
\code{download_OISST()} is a function to download OISST v2.1

If the number of cores is set to >1, the function uses \code{mclapply} to parallel download datasets, else 1 = single downloads.

Notes:
The NOAA 1/4° Daily Optimum Interpolation Sea Surface Temperature (OISST) is a long term Climate Data Record that incorporates
observations from different platforms (satellites, ships, buoys and Argo floats) into a regular global grid. The dataset is
interpolated to fill gaps on the grid and create a spatially complete map of sea surface temperature. Satellite and ship observations
are referenced to buoys to compensate for platform differences and sensor biases.
}
\examples{
\dontrun{

download_OISST(url = "https://www.ncei.noaa.gov/data/sea-surface-temperature-optimum-interpolation/v2.1/access/avhrr/",
  start_date = "1990-01-01",
  end_date = "1990-01-02",
  dest_dir = "/Volumes/Extreme_SSD/dhw/OISST/2024/",
  mc.cores = 1)

download_OISST(url = "https://www.ncei.noaa.gov/data/sea-surface-temperature-optimum-interpolation/v2.1/access/avhrr/",
  dates = c("1990-01-01", "1990-01-02")
  dest_dir = "/Volumes/Extreme_SSD/dhw/OISST/2024/",
  mc.cores = 1)

}
}
