% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_reefs.R
\name{extract_reefs}
\alias{extract_reefs}
\title{Extract reef}
\usage{
extract_reefs(
  input,
  output = "sf",
  shpfile,
  extract_fun = "weighted_mean",
  quiet = FALSE
)
}
\arguments{
\item{input}{input}

\item{output}{output format, sf default}

\item{shpfile}{location of shpfile mask}

\item{extract_fun}{= "weighted_mean" by default}

\item{quiet}{show verbose}
}
\value{
shp file sf with SST details (see above for details)
}
\description{
Function to extract sst data for shp file overlay

See vignette for further details
}
\examples{
\dontrun{
output <- extract_reefs()
}
}
