% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_annual_dhw.R, R/plot_mm.R
\name{plot_mm}
\alias{plot_mm}
\alias{plot_annual_DHW}
\title{Plot mean monthly maximum for a given gridcell}
\usage{
plot_annual_DHW(input, fixedyears = NULL)

plot_mm(input, lon, lat)
}
\arguments{
\item{input}{rast() file}

\item{fixedyears}{range y lims for ggplot}

\item{lon}{vector of longitude}

\item{lat}{vector of latitude}
}
\value{
ggplot of annual DHW scaled Nov-Aug

ggplot of baseline climatology
}
\description{
Plot mean monthly maximum for a given gridcell
from a terra::rast() file input

Plot mean monthly maximum for a given gridcell
from a terra::rast() file input
}
