% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ERA5.R
\name{process_ERA5}
\alias{process_ERA5}
\title{Process ERA5 NetCDF Files}
\usage{
process_ERA5(
  input,
  polygon,
  crop = TRUE,
  mask = TRUE,
  downsample = FALSE,
  res = 0.1,
  crs = "EPSG:7844",
  combinedfilename = NULL,
  units = "celsius"
)
}
\arguments{
\item{input}{Folder containing NetCDF (.nc) files.}

\item{polygon}{polygon for crop/masl}

\item{crop}{TRUE/FALSE}

\item{mask}{TRUE/FALSE}

\item{downsample}{TRUE/FALSE}

\item{res}{resolution for downsamlpling}

\item{crs}{change the CRS if needed (EPSG:4283 as default)}

\item{combinedfilename}{output file path, should be .rds}

\item{units}{Units for temperature: one of "celsius" or "kelvin". Default is "celsius".}

\item{variable}{redundant?}
}
\value{
A combined SpatRaster object in ERA5 format.
}
\description{
Function to extract and combine SST data from NetCDF files in a specified folder.

See vignette for further details.
}
\examples{
\dontrun{
folder <- "/Users/rof011/GBR-dhw/datasets/era5/"
process_ERA5(folder, units = "celsius")
}
}
